experiments/two_moons/requirements_sbi.txt
absl-py==1.4.0
altair==5.0.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arviz==0.15.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autograd==1.5
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
dask==2023.5.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deneb==1.0.1
diffeqtorch==1.0.0
distributed==2023.5.0
elfi==0.8.6
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
flatlatex==0.15
fonttools==4.39.4
fqdn==1.5.1
fsspec==2023.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
GPy==1.10.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5netcdf==1.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.0
ipyparallel==8.6.1
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jabbar==0.0.15
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
julia==0.6.1
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mergedeep==1.3.4
mistune==2.0.5
mpmath==1.2.1
msgpack==1.0.5
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nflows==0.14
notebook==6.5.4
notebook_shim==0.2.3
numdifftools==0.9.41
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
paramz==0.9.5
parso==0.8.3
partd==1.4.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
portpicker==1.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyabc==0.12.10
pyabcranger==0.0.69
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyknos==0.15.1
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.5
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
sbi==0.21.0
sbibm==1.0.8
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
sympy==1.11.1
tblib==1.7.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.13
vega-datasets==0.9.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
wsproto==1.2.0
xarray==2023.4.2
xarray-einstats==0.5.1
zict==3.0.0
zipp==3.15.0
